+++ lib/http_base.ex
@@ -5,8 +5,18 @@ defmodule Astra.HttpBase do
   def user_agent() do
     "AstraEx/#{@version}"
   end
+  
+  # Allow for either Astra regions and id or direct stargate urls
+  def url_base() do
+    config = Application.get_all_env(:astra)
+    find_base(config, Keyword.has_key?(config, :stargate_url))
+  end
+  
+  def find_base(config, true), do: config[:url]
+  def find_base(config, false), do: "https://#{config[:id]}-#{config[:region]}.apps.astra.datastax.com/api/"
     
-  defmacro __using__(opts) do  
+  defmacro __using__(opts) do
+    # the sub-path for the service we are accessing, this is added to the url_base
     path = Keyword.get(opts, :path, "")
 
     quote do
@@ -25,9 +35,7 @@ defmodule Astra.HttpBase do
       end
       
       def process_request_url(sub_path) do
-        config = Application.get_all_env(:astra)
-        url = "https://#{config[:id]}-#{config[:region]}.apps.astra.datastax.com/api/"
-        url <> unquote(path) <> sub_path
+        Astra.HttpBase.url_base() <> unquote(path) <> sub_path
       end
       
       def process_request_headers(headers) do